+++ quint/src/repl.ts
@@ -11,17 +11,17 @@
 import * as readline from 'readline'
 import { Readable, Writable } from 'stream'
 import { readFileSync, writeFileSync } from 'fs'
-import { cwd } from 'process'
 import lineColumn from 'line-column'
 import { Maybe, just, none } from '@sweet-monads/maybe'
 import { Either, left, right } from '@sweet-monads/either'
 import chalk from 'chalk'
 import { format } from './prettierimp'
 
-import { QuintEx } from './quintIr'
+import { FlatDef, QuintEx, isFlat } from './quintIr'
 import {
   CompilationContext,
   CompilationState,
+  compileDef,
   compileExpr,
   compileFromCode,
   contextNameLookup,
@@ -31,13 +31,15 @@ import {
 import { formatError } from './errorReporter'
 import { ComputableKind, EvalResult, Register, kindName } from './runtime/runtime'
 import { TraceRecorder, newTraceRecorder, noExecutionListener } from './runtime/trace'
-import { ErrorMessage, parseExpressionOrUnit } from './quintParserFrontend'
+import { ErrorMessage, parseExpressionOrUnit } from './parsing/quintParserFrontend'
 import { prettyQuintEx, printExecutionFrameRec, terminalWidth } from './graphics'
 import { verbosity } from './verbosity'
 import { newRng } from './rng'
 import { version } from './version'
-import { fileSourceResolver } from './sourceResolver'
+import { fileSourceResolver } from './parsing/sourceResolver'
+import { cwd } from 'process'
 import { newIdGenerator } from './idGenerator'
+import { definitionToString } from './IRprinting'
 
 // tunable settings
 export const settings = {
@@ -121,8 +123,14 @@ export function quintRepl(
   function replInitialCompilationState(): CompilationState {
     // Introduce the __repl__ module to the compilation state so new expressions
     // are pushed to it.
-    return { ...newCompilationState(), modules: [{ name: '__repl__', defs: [], id: 0n }] }
+    const state = newCompilationState()
+    return { ...state, modules: [{ name: '__repl__', defs: simulatorBuiltins(state), id: 0n }] }
   }
+
+  // Add the builtins to the text history as well
+  state.defsHist = state.compilationState.modules[0].defs.map(def => definitionToString(def)).join('\n')
+
+  tryEvalHistory(out, state)
   // we let the user type a multiline string, which is collected here:
   let multilineText = ''
   // when recyclingOwnOutput is true, REPL is receiving its older output
@@ -383,14 +391,10 @@ function loadFromFile(out: writer, state: ReplState, filename: string): boolean 
 
       // save the definition history
       newState.defsHist = defsAndExprs[0]
-
-      const isError = !tryEval(out, newState, 'val __loaded = true')
+      const isError = !tryEvalHistory(out, newState)
       if (isError) {
         return false
       }
-
-      // reset the definiton history to the one loaded from the file
-      newState.defsHist = defsAndExprs[0]
 
       if (defsAndExprs.length > 1) {
         // unwrap the expressions from the specially crafted comments
@@ -402,6 +406,11 @@ function loadFromFile(out: writer, state: ReplState, filename: string): boolean 
             return false
           }
         }
+      }
+    } else {
+      const isError = !tryEvalHistory(out, newState)
+      if (isError) {
+        return false
       }
     }
 
@@ -478,41 +487,65 @@ function loadShadowVars(state: ReplState, context: CompilationContext): void {
 
 // Declarations that are overloaded by the simulator.
 // In the future, we will declare them in a separate module.
-const simulatorBuiltins = `val ${lastTraceName} = [];
-def q::test(q::nruns, q::nsteps, q::init, q::next, q::inv) = false;
-def q::testOnce(q::nsteps, q::init, q::next, q::inv) = false;
-`
+function simulatorBuiltins(compilationState: CompilationState): FlatDef[] {
+  return [
+    parseDefOrThrow(compilationState, `val ${lastTraceName} = []`),
+    parseDefOrThrow(compilationState, `def q::test = (q::nruns, q::nsteps, q::init, q::next, q::inv) => false`),
+    parseDefOrThrow(compilationState, `def q::testOnce = (q::nsteps, q::init, q::next, q::inv) => false`),
+  ]
+}
 
-// try to evaluate the expression in a string and print it, if successful
-function tryEval(out: writer, state: ReplState, newInput: string): boolean {
-  // output errors to the console in red
-  function printErrors(moduleText: string, context: CompilationContext) {
-    printErrorMessages(out, 'syntax error', newInput, moduleText, context.syntaxErrors)
-    printErrorMessages(out, 'static analysis error', newInput, moduleText, context.analysisErrors, chalk.yellow)
-    printErrorMessages(out, 'compile error', moduleText, newInput, context.compileErrors)
-    out('\n') // be nice to external programs
+function parseDefOrThrow(compilationState: CompilationState, text: string): FlatDef {
+  const result = parseExpressionOrUnit(text, '<builtins>', compilationState.idGen, compilationState.sourceMap)
+  if (result.kind === 'toplevel' && isFlat(result.def)) {
+    return result.def
+  } else {
+    throw new Error(`Expected a flat definition, got ${result.kind}, parsing ${text}`)
   }
+}
 
-  // Compose the input to the parser.
-  // TODO: REPL should work incrementally:
-  // https://github.com/informalsystems/quint/issues/618
-  function prepareParserInput(textToAdd: string): string {
-    return `${state.moduleHist}
-module __repl__ { ${simulatorBuiltins}
+// Compose the text input. This is not evaluated, it's only used to
+// report errors, as the error locs are relative to the input.
+function prepareModulesText(state: ReplState, textToAdd: string): string {
+  return `${state.moduleHist}
+module __repl__ {
 ${state.defsHist}
 ${textToAdd}
 }`
+}
+
+function tryEvalHistory(out: writer, state: ReplState): boolean {
+  const modulesText = prepareModulesText(state, '')
+  const mainPath = fileSourceResolver().lookupPath(cwd(), 'repl.ts')
+  const rng = newRng(state.seed)
+
+  const context = compileFromCode(newIdGenerator(), modulesText, '__repl__', mainPath, noExecutionListener, rng.next)
+  if (context.values.size === 0 || context.compileErrors.length > 0 || context.syntaxErrors.length > 0) {
+    printErrors(out, state, context)
+    return false
   }
 
+  if (context.analysisErrors.length > 0) {
+    printErrors(out, state, context)
+    // provisionally, continue on type & effects errors
+  }
+
+  // Save compilation state
+  state.compilationState = context.compilationState
+
+  return true
+}
+
+// try to evaluate the expression in a string and print it, if successful
+function tryEval(out: writer, state: ReplState, newInput: string): boolean {
   const parseResult = parseExpressionOrUnit(
     newInput,
     '<input>',
     state.compilationState.idGen,
     state.compilationState.sourceMap
   )
-  const moduleText = prepareParserInput(newInput)
   if (parseResult.kind === 'error') {
-    printErrorMessages(out, 'syntax error', newInput, moduleText, parseResult.messages)
+    printErrorMessages(out, state, 'syntax error', newInput, parseResult.messages)
     out('\n') // be nice to external programs
     return false
   }
@@ -528,7 +561,7 @@ ${textToAdd}
     const context = compileExpr(state.compilationState, rng, recorder, parseResult.expr)
 
     if (context.syntaxErrors.length > 0 || context.compileErrors.length > 0 || context.analysisErrors.length > 0) {
-      printErrors(newInput, context)
+      printErrors(out, state, context, newInput)
       if (context.syntaxErrors.length > 0 || context.compileErrors.length > 0) {
         return false
       } // else: provisionally, continue on type & effects errors
@@ -540,7 +573,7 @@ ${textToAdd}
     return evalExpr(state, context, recorder, out)
       .mapLeft(msg => {
         // when #618 is implemented, we should remove this
-        printErrorMessages(out, 'runtime error', newInput, moduleText, context.getRuntimeErrors())
+        printErrorMessages(out, state, 'runtime error', newInput, context.getRuntimeErrors())
         // print the error message produced by the lookup
         out(chalk.red(msg))
         out('\n') // be nice to external programs
@@ -549,19 +582,16 @@ ${textToAdd}
   }
   if (parseResult.kind === 'toplevel') {
     // compile the module and add it to history if everything worked
-    const mainPath = fileSourceResolver().lookupPath(cwd(), 'repl.ts')
 
-    // For toplevel definitions, we start from scratch. This should be made
-    // incremental as well.
-    const context = compileFromCode(newIdGenerator(), moduleText, '__repl__', mainPath, noExecutionListener, rng.next)
+    const context = compileDef(state.compilationState, rng, noExecutionListener, parseResult.def)
 
     if (context.values.size === 0 || context.compileErrors.length > 0 || context.syntaxErrors.length > 0) {
-      printErrors(moduleText, context)
+      printErrors(out, state, context, newInput)
       return false
     }
 
     if (context.analysisErrors.length > 0) {
-      printErrors(moduleText, context)
+      printErrors(out, state, context, newInput)
       // provisionally, continue on type & effects errors
     }
 
@@ -575,18 +605,27 @@ ${textToAdd}
   return true
 }
 
+// output errors to the console in red
+function printErrors(out: writer, state: ReplState, context: CompilationContext, newInput: string = '') {
+  printErrorMessages(out, state, 'syntax error', newInput, context.syntaxErrors)
+  printErrorMessages(out, state, 'static analysis error', newInput, context.analysisErrors, chalk.yellow)
+  printErrorMessages(out, state, 'compile error', newInput, context.compileErrors)
+  out('\n') // be nice to external programs
+}
+
 // print error messages with proper colors
 function printErrorMessages(
   out: writer,
+  state: ReplState,
   kind: string,
   inputText: string,
-  moduleText: string,
   messages: ErrorMessage[],
   color: (_text: string) => string = chalk.red
 ) {
+  const modulesText = prepareModulesText(state, inputText)
   // display the error messages and highlight the error places
   messages.forEach(e => {
-    const text = e.locs[0].source === '<input>' ? inputText : moduleText
+    const text = e.locs[0].source === '<input>' ? inputText : modulesText
     const finder = lineColumn(text)
     const msg = formatError(text, finder, e, none())
     out(color(`${kind}: ${msg}\n`))